requirements.txt
fastapi
uvicorn[standard]
aleph-alpha-client
python-dotenv
timm
transformers
opencv-python
python-multipart